added scrtipt o update server
